[core] move platform/gl.hpp to gl/gl.hpp
