+++ platform/android/src/native_map_view.cpp
@@ -15,7 +15,7 @@
 #include <mbgl/platform/platform.hpp>
 #include <mbgl/platform/event.hpp>
 #include <mbgl/platform/log.hpp>
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 #include <mbgl/util/constants.hpp>
 
 namespace mbgl {
