+++ platform/osx/src/MGLMapView.mm
@@ -18,7 +18,7 @@
 #import <mbgl/annotation/point_annotation.hpp>
 #import <mbgl/map/camera.hpp>
 #import <mbgl/platform/darwin/reachability.h>
-#import <mbgl/platform/gl.hpp>
+#import <mbgl/gl/gl.hpp>
 #import <mbgl/sprite/sprite_image.hpp>
 #import <mbgl/storage/default_file_source.hpp>
 #import <mbgl/storage/network_status.hpp>
