+++ platform/osx/src/MGLOpenGLLayer.mm
@@ -2,7 +2,7 @@
 
 #import "MGLMapView_Private.h"
 
-#import <mbgl/platform/gl.hpp>
+#import <mbgl/gl/gl.hpp>
 
 @implementation MGLOpenGLLayer
