+++ src/mbgl/geometry/fill_buffer.cpp
@@ -1,6 +1,6 @@
 #include <mbgl/geometry/fill_buffer.hpp>
 
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 
 #include <climits>
