+++ src/mbgl/geometry/glyph_atlas.cpp
@@ -2,7 +2,7 @@
 
 #include <mbgl/text/font_stack.hpp>
 
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 #include <mbgl/platform/log.hpp>
 #include <mbgl/platform/platform.hpp>
 #include <mbgl/util/gl_object_store.hpp>
