+++ src/mbgl/geometry/line_atlas.hpp
@@ -1,7 +1,7 @@
 #ifndef MBGL_GEOMETRY_LINE_ATLAS
 #define MBGL_GEOMETRY_LINE_ATLAS
 
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 
 #include <vector>
 #include <map>
