+++ src/mbgl/gl/debugging.cpp
@@ -1,5 +1,5 @@
 #include <mbgl/gl/debugging.hpp>
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 #include <mbgl/platform/event.hpp>
 #include <mbgl/platform/log.hpp>
