+++ src/mbgl/renderer/painter.hpp
@@ -14,7 +14,7 @@
 
 #include <mbgl/style/types.hpp>
 
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 
 #include <mbgl/util/noncopyable.hpp>
 #include <mbgl/util/chrono.hpp>
