+++ src/mbgl/shader/icon_shader.cpp
@@ -1,7 +1,7 @@
 #include <mbgl/shader/icon_shader.hpp>
 #include <mbgl/shader/icon.vertex.hpp>
 #include <mbgl/shader/icon.fragment.hpp>
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 
 #include <cstdio>
