+++ src/mbgl/shader/shader.hpp
@@ -1,7 +1,7 @@
 #ifndef MBGL_RENDERER_SHADER
 #define MBGL_RENDERER_SHADER
 
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 #include <mbgl/util/noncopyable.hpp>
 
 #include <cstdint>
