+++ src/mbgl/util/raster.cpp
@@ -1,5 +1,5 @@
 #include <mbgl/platform/platform.hpp>
-#include <mbgl/platform/gl.hpp>
+#include <mbgl/gl/gl.hpp>
 #include <mbgl/platform/log.hpp>
 
 #include <mbgl/util/raster.hpp>
